more refactoring -untested
