+++ package.json
@@ -18,6 +18,7 @@
     "body-parser": "^1.17.1",
     "connect-ensure-login": "^0.1.1",
     "cookie-parser": "~1.3.5",
+    "crypto": "0.0.3",
     "express": "^4.15.2",
     "express-session": "^1.11.3",
     "forms": "^1.3.0",
